# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/api/login/": {
            "post": {
                "summary": "login",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/logout/": {
            "post": {
                "summary": "User logout",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "jwt token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "api/check-token/{user_id}/{token}": {
            "post": {
                "summary": "Check User Token",
                "operationId": "token",
                "parameters": [
                    {
                        "name": "user Id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "path",
                        "description": "jwt token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/register/": {
            "get": {
                "summary": "Registeration Form View",
                "operationId": "register",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "summary": "Add new user",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "facutly",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "university",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "DOB",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "testing data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "testing data",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "committee",
                        "in": "query",
                        "description": "testing data",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ex_options",
                        "in": "query",
                        "description": "testing data",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "summary": "User Profile",
                "operationId": "profile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": "Bearer",
                        "type": "string",
                        "format": "JWT"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {}
}